requirements.txt
bidict==0.22.0
pyserial==3.5
python-engineio==4.3.4
python-socketio==5.7.2
threaded==4.1.0
tornado==6.2
